requirements.txt
pandas==1.5.2
numpy==1.24.0
streamlit==1.16.0
google-cloud==0.34.0
google-auth==2.15.0
matplotlib==3.6.2
altair==4.2.0